+++ src/game.rs
@@ -189,10 +189,9 @@ impl<S: System> Game<S> for PixcilGame {
             "importImage" => {
                 let image = decode_sprite(data).or_fail()?;
                 let event = Event::Import { image };
-                self.handle_pixcil_event(system, Some(event)).or_fail()?;
-
                 let app = self.app.as_mut().or_fail()?;
                 app.request_redraw(app.screen_size().to_region());
+                self.handle_pixcil_event(system, Some(event)).or_fail()?;
                 Ok(())
             }
             "disableSaveWorkspaceButton" => {