+++ src/io.rs
@@ -6,6 +6,7 @@ pub enum IoRequest {
     SaveWorkspace,
     LoadWorkspace,
     ImportImage,
+    ImportImageFromClipboard,
     InputNumber { id: InputId },
     Vibrate,
 }